Doc_Project/requirements.txt
python-docx==0.8.11
nltk==3.8.1
pandas==2.0.3
openpyxl==3.1.2
markdown==3.5.2
html2text==2020.1.16 